[TASK] Implemented the IsRunning function for RunProviders.
This makes it easy to see if a Provider is running, both from the Stack and from other Providers that depend on it.
Also moved the name() function used within the Stack to a separate utility file within the provider namespace.
Also added a "WaitForRunningProvider" function to this utility file, which allows a class/provider to wait until a Provider is running.
That saves a lot of potential nil pointer panics.