[TASK] PIMSAA-57: Added middlewares.
These can be use to add interceptors to the HTTP handlers.
Currently, only the GraphQL provider supports them and only the JWTMiddleware exists.

This JWTMiddleware checks and validates a JWT token from the request and sets the token as context value.
This token can be used all across the application, for example to send it to other services, or to use claims.