PIM-118: Added a custom JsonPbMarshaller that uses the gogo-gateway marshaller.
Originally, the logrus payload interceptors for GRPC messages use the standard golang marshaller, but this doesn't properly support the gogo types of messages.
To fix this, I've created a PR in the grpc-middleware project (see github.com/grpc-ecosystem/go-grpc-middleware/pull/236), so we can override the marshaller with our own custom one.
The replace in go.mod is because of this PR, this can be removed once grpc-middleware gets a new release.